Refactor CommentHighlighter to supported language specific types
